MODSOURCE-562 - refactored
